Addition of early stages of !subr.
This commit adds a very basic, simple command "!subr" with the intent of fetching things from a user given subreddit. The command will only now return a random post from (I believe only the first page) of the "hot" section of the subreddit. This will be expanded in the future.